:white_check_mark: add create comment test
