+++ social-media-react/src/components/posts/__tests__/UpdatePost.test.js
@@ -1,7 +1,6 @@
 import { render, screen, fireEvent } from "../../../helpers/test-utils";
 import userEvent from "@testing-library/user-event";
 import UpdatePost from "../UpdatePost";
-import { setUserData } from "../../../hooks/user.actions";
 import userFixtures from "../../../helpers/fixtures/user";
 import postFixtures from "../../../helpers/fixtures/post";
 import { faker } from "@faker-js/faker";
@@ -9,19 +8,6 @@ import { faker } from "@faker-js/faker";
 const user = userFixtures();
 
 const post = postFixtures(true, false, user);
-
-beforeEach(() => {
-  // to fully reset the state between __tests__, clear the storage
-  localStorage.clear();
-  // and reset all mocks
-  jest.clearAllMocks();
-
-  setUserData({
-    user: user,
-    access: null,
-    refresh: null,
-  });
-});
 
 test("Create Post form renders", async () => {
   const user = userEvent.setup();
